feat: add anti-fraud audit trail for overridden AI detections

+++ src/main/java/com/scrapDetection/dto/detection/DetectionRequestDTO.java
@@ -64,10 +64,10 @@ public class DetectionRequestDTO {
         @Setter
         @NoArgsConstructor
         public static class BBox {
-            private Integer x1;
-            private Integer y1;
-            private Integer x2;
-            private Integer y2;
+            private Double x1;
+            private Double y1;
+            private Double x2;
+            private Double y2;
         }
     }
 }
