+++ src/main/java/com/scrapDetection/service/impl/DetectionServiceImpl.java
@@ -98,10 +98,10 @@ public class DetectionServiceImpl implements DetectionService {
 
         // Extract bbox from the best detection (may be null)
         DetectionItemDTO.BBox bbox = best.getBbox();
-        Integer bx1 = bbox != null ? bbox.getX1() : null;
-        Integer by1 = bbox != null ? bbox.getY1() : null;
-        Integer bx2 = bbox != null ? bbox.getX2() : null;
-        Integer by2 = bbox != null ? bbox.getY2() : null;
+        Double bx1 = bbox != null ? bbox.getX1() : null;
+        Double by1 = bbox != null ? bbox.getY1() : null;
+        Double bx2 = bbox != null ? bbox.getX2() : null;
+        Double by2 = bbox != null ? bbox.getY2() : null;
 
         // ── 4. Build response, stamp timestamp, store as "latest" ──────────────
         DetectionResponseDTO response = detectionMapper.toReceivedResponse(
